requirements.txt
Flask==2.2.2
langchain==0.1.13
langchain_community==0.0.29
langchain_experimental==0.0.55
langchain_google_genai==1.0.1
PyMySQL==1.1.0
python-dotenv==1.0.1
werkzeug==2.2.2
chromadb
sentence-transformers==2.3.1